# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "simple description",
        "title": "QrMenuApi",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/company/{companyID}/pubs": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get all pubs for company",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pub"
                ],
                "summary": "Get all pubs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company id",
                        "name": "companyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pubs.getAllPubsForCompanyOutput"
                        }
                    }
                }
            }
        },
        "/company/{companyID}/pubs/": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates pub",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pub"
                ],
                "summary": "Create pub",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company id",
                        "name": "companyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "pub params",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.CreatePubInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/pubs.cratePubOutput"
                        }
                    }
                }
            }
        },
...
        "/company/{companyID}/pubs/{pubID}/menus/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Gets all menus for pub",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "Get all menu",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company id",
                        "name": "companyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pub id",
                        "name": "pubID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/menus.createMenuOutput"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates menu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "Create menu",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company id",
                        "name": "companyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pub id",
                        "name": "pubID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "menu params",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.MenuInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/menus.createMenuOutput"
                        }
                    }
                }
            }
        },
        "/company/{companyID}/pubs/{pubID}/menus/{menuID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Gets info about menu",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "Get menu",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company id",
                        "name": "companyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pub id",
                        "name": "pubID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "menu id",
                        "name": "menuID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/menus.getMenuOutput"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates menu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "Update menu",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company id",
                        "name": "companyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pub id",
                        "name": "pubID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "menu id",
                        "name": "menuID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "menu params",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.MenuInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/menus.updateMenuOutput"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deletes menu",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "menu"
                ],
                "summary": "Delete menu",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "company id",
                        "name": "companyID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "pub id",
                        "name": "pubID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "menu id",
                        "name": "menuID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "access_token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/menus.delteMenuOutput"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entities.CreatePubInput": {
            "type": "object",
            "required": [
                "name",
                "url_name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "minLength": 2,
                    "example": "My pub name"
                },
                "url_name": {
                    "type": "string",
                    "minLength": 2,
                    "example": "my-pub-name"
                }
            }
        },
...
        "entities.MenuInput": {
            "type": "object",
            "required": [
                "name",
                "pub_id",
                "visible"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "minLength": 2,
                    "example": "My menu name"
                },
                "pub_id": {
                    "type": "integer",
                    "example": 2
                },
                "visible": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "entities.MenuOutput": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "My menu name"
                },
                "pub_id": {
                    "type": "integer",
                    "example": 2
                },
                "visible": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "entities.PubOutput": {
            "type": "object",
            "properties": {
                "additional_info": {
                    "type": "string",
                    "example": "My pub additional info"
                },
                "address": {
                    "type": "string",
                    "example": "My pub address"
                },
                "bg_image_file_name": {
                    "type": "string",
                    "example": "my-pub-name-bg.png"
                },
                "color": {
                    "type": "string",
                    "example": "#ffffff"
                },
                "color_theme": {
                    "type": "string",
                    "example": "light"
                },
                "company_id": {
                    "type": "integer"
                },
                "currency_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "logo_file_name": {
                    "type": "string",
                    "example": "my-pub-name-logo.png"
                },
                "name": {
                    "type": "string",
                    "example": "My pub name"
                },
                "qr_code_file_name": {
                    "type": "string",
                    "example": "my-pub-name.png"
                },
                "url_name": {
                    "type": "string",
                    "example": "my-pub-name"
                },
                "wifi_password": {
                    "type": "string",
                    "example": "12345678"
                }
            }
        },
...
        "menus.createMenuOutput": {
            "type": "object",
            "properties": {
                "menu": {
                    "$ref": "#/definitions/entities.MenuOutput"
                },
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "menus.delteMenuOutput": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "menus.getMenuOutput": {
            "type": "object",
            "properties": {
                "menu": {
                    "$ref": "#/definitions/entities.MenuOutput"
                },
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "menus.updateMenuOutput": {
            "type": "object",
            "properties": {
                "menu": {
                    "$ref": "#/definitions/entities.MenuOutput"
                },
                "ok": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "pubs.cratePubOutput": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean",
                    "example": true
                },
                "pub": {
                    "$ref": "#/definitions/entities.PubOutput"
                }
            }
        },
...
        "pubs.getAllPubsForCompanyOutput": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean",
                    "example": true
                },
                "pubs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.PubOutput"
                    }
                }
            }
        },
...
    }
}