# src/controllers/httpv1/routes/pubs/create-pub.go
package pubs

import (
	"strconv"

	h "github.com/alexkalak/qrmenu/src/controllers/httpv1/httphelpers"
	"github.com/alexkalak/qrmenu/src/controllers/httpv1/input"
	"github.com/alexkalak/qrmenu/src/controllers/httpv1/input/entities"
	"github.com/alexkalak/qrmenu/src/errors/httperrors"
	"github.com/gofiber/fiber/v2"
)

type cratePubOutput struct {
	Ok  bool               `json:"ok" example:"true"`
	Pub entities.PubOutput `json:"pub"`
}

// @Summary      Create pub
// @Description  Creates pub
// @Tags         pub
// @Param companyID path int true "company id"
// @Param input body entities.CreatePubInput true "pub params"
// @Accept       json
// @Produce      json
// @Success      201  {object}  cratePubOutput
// @Router       /company/{companyID}/pubs/ [POST]
// @Security ApiKeyAuth
// @Param access_token header string  true "access_token"
func (c *pubController) CreatePub(ctx *fiber.Ctx) error {
	userID, userSignificance, err := h.GetUserIDAndSignificanceFromLocals(ctx)
	if err != nil {
		return h.SendError(ctx, err, h.AUTOMATIC_STATUS_CODE)
	}

	companyID, err := strconv.Atoi(ctx.Params("companyID"))
	if err != nil {
		return h.SendError(ctx, httperrors.ErrBadID, h.AUTOMATIC_STATUS_CODE)
	}

	err = h.CheckAccess(userID, companyID, userSignificance)
	if err != nil {
		return h.SendError(ctx, err, h.AUTOMATIC_STATUS_CODE)
	}

	input, validationErrors, err := input.ParseRequestBody[entities.CreatePubInput](ctx)
	if err != nil {
		return h.SendError(ctx, err, h.AUTOMATIC_STATUS_CODE)
	}
	if len(validationErrors) > 0 {
		return h.SendValidationErrors(ctx, validationErrors)
	}

	inputPub := input.ConvertToModel(companyID)
	pub, err := c.PubService.CreatePub(inputPub)
	if err != nil {
		return h.SendError(ctx, err, h.AUTOMATIC_STATUS_CODE)
	}

	output := entities.PubOutput{}
	output.ConvertFromModel(pub)

	return h.SendSuccess(
		ctx,
		fiber.Map{
			"pub": output,
		},
		fiber.StatusCreated)
}